쇼츠생성기뉴/backend/requirements.txt
flask
flask-cors
google-generativeai
openai
python-dotenv
requests
moviepy
edge-tts
pillow
numpy
proglog
